Refactors kanaria config on tests to its own script.
